requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=0.6,!=0.7,<1.0
Babel>=1.3
pika>=0.9
tornado>=4.0
sockjs-tornado>=1.0
eventlet>=0.15.2
six>=1.7.0
